requirements.txt
tkinter
requests
fal-client
Pillow